[fix] Fix debug drawing in BaseBoneSpace
The debug drawing for bones and collisions was not being displayed correctly when the SimulationSpace was set to BaseBoneSpace.

This was because the data in the RuntimeNode was in a coordinate system based on the SimulationSpace, but it was being rendered directly as if it were in component space.

This commit fixes the issue by transforming the bone and collision data to component space before rendering, only when the SimulationSpace is BaseBone-based.