[fix] Fix text position of LengthRateFromRoot in BaseBoneSpace
The text for LengthRateFromRoot was not being displayed in the correct position when the SimulationSpace was set to BaseBoneSpace.

This was because the bone location used for the text position was not being transformed from BaseBoneSpace to component space.

This commit fixes the issue by transforming the bone location to component space before calculating the final text position.